+++ plugins/catalog/src/components/CatalogTable/PaginatedCatalogTable.tsx
@@ -18,10 +18,7 @@ import React from 'react';
 
 import { Table, TableProps } from '@backstage/core-components';
 import { CatalogTableRow } from './types';
-import {
-  EntityTextFilter,
-  useEntityList,
-} from '@backstage/plugin-catalog-react';
+import { CatalogTableToolbar } from './CatalogTableToolbar';
 
 type PaginatedCatalogTableProps = {
   prev?(): void;
@@ -33,7 +30,6 @@ type PaginatedCatalogTableProps = {
  */
 export function PaginatedCatalogTable(props: PaginatedCatalogTableProps) {
   const { columns, data, next, prev, title, isLoading, options } = props;
-  const { updateFilters } = useEntityList();
 
   return (
     <Table
@@ -49,17 +45,15 @@ export function PaginatedCatalogTable(props: PaginatedCatalogTableProps) {
         pageSize: Number.MAX_SAFE_INTEGER,
         emptyRowsWhenPaging: false,
       }}
-      onSearchChange={(searchText: string) =>
-        updateFilters({
-          text: searchText ? new EntityTextFilter(searchText) : undefined,
-        })
-      }
       onPageChange={page => {
         if (page > 0) {
           next?.();
         } else {
           prev?.();
         }
+      }}
+      components={{
+        Toolbar: CatalogTableToolbar,
       }}
       /* this will enable the prev button accordingly */
       page={prev ? 1 : 0}