+++ plugins/github-issues/src/components/GithubIssues/GithubIssues.tsx
@@ -15,7 +15,9 @@
  */
 
 import React from 'react';
-import { Box, IconButton, Typography } from '@material-ui/core';
+import Box from '@material-ui/core/Box';
+import IconButton from '@material-ui/core/IconButton';
+import Typography from '@material-ui/core/Typography';
 import { InfoCard, Progress } from '@backstage/core-components';
 import RefreshIcon from '@material-ui/icons/Refresh';
 import { useEntityGithubRepositories } from '../../hooks/useEntityGithubRepositories';