+++ packages/backend-tasks/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @backstage/backend-tasks
 
+## 0.5.20
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-common@0.21.5
+  - @backstage/config@1.2.0
+  - @backstage/errors@1.2.4
+  - @backstage/types@1.1.1
+
 ## 0.5.19
 
 ### Patch Changes