+++ packages/dev-utils/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @backstage/dev-utils
 
+## 1.0.29
+
+### Patch Changes
+
+- e8f026a: Use ESM exports of react-use library
+- Updated dependencies
+  - @backstage/core-components@0.14.2
+  - @backstage/core-app-api@1.12.2
+  - @backstage/plugin-catalog-react@1.11.1
+  - @backstage/app-defaults@1.5.2
+  - @backstage/integration-react@1.1.25
+  - @backstage/core-plugin-api@1.9.1
+  - @backstage/catalog-model@1.4.5
+  - @backstage/theme@0.5.2
+
 ## 1.0.28
 
 ### Patch Changes