+++ plugins/entity-feedback/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @backstage/plugin-entity-feedback
 
+## 0.2.15
+
+### Patch Changes
+
+- 12610b4: Added ESLint rule `no-top-level-material-ui-4-imports` in the `entity-feedback` plugin to migrate the Material UI imports.
+- e8f026a: Use ESM exports of react-use library
+- Updated dependencies
+  - @backstage/core-components@0.14.2
+  - @backstage/plugin-catalog-react@1.11.1
+  - @backstage/core-plugin-api@1.9.1
+  - @backstage/catalog-model@1.4.5
+  - @backstage/errors@1.2.4
+  - @backstage/plugin-entity-feedback-common@0.1.3
+
 ## 0.2.14
 
 ### Patch Changes