+++ plugins/sonarqube/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @backstage/plugin-sonarqube
 
+## 0.7.14
+
+### Patch Changes
+
+- e8f026a: Use ESM exports of react-use library
+- Updated dependencies
+  - @backstage/core-components@0.14.2
+  - @backstage/plugin-catalog-react@1.11.1
+  - @backstage/core-plugin-api@1.9.1
+  - @backstage/catalog-model@1.4.5
+  - @backstage/plugin-sonarqube-react@0.1.14
+
 ## 0.7.13
 
 ### Patch Changes