+++ plugins/techdocs-addons-test-utils/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @backstage/plugin-techdocs-addons-test-utils
 
+## 1.0.29
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-techdocs-react@1.2.1
+  - @backstage/core-app-api@1.12.2
+  - @backstage/plugin-catalog-react@1.11.1
+  - @backstage/plugin-search-react@1.7.8
+  - @backstage/plugin-techdocs@1.10.2
+  - @backstage/plugin-catalog@1.18.1
+  - @backstage/integration-react@1.1.25
+  - @backstage/core-plugin-api@1.9.1
+  - @backstage/test-utils@1.5.2
+
 ## 1.0.28
 
 ### Patch Changes